@@ -1,7 +1,10 @@
 name: chatverse
 description: A comprehensive Flutter chat library with Firebase integration, featuring advanced UI, group chat support, and real-time messaging capabilities.
-version: 0.0.1
-homepage: https://github.com/yourusername/chatverse
+version: 0.0.4
+homepage: https://github.com/Hamed233/ChatVerse
+repository: https://github.com/Hamed233/ChatVerse
+issue_tracker: https://github.com/Hamed233/ChatVerse/issues
+documentation: https://github.com/Hamed233/ChatVerse/wiki
 
 environment:
   sdk: '>=3.0.0 <4.0.0'
@@ -10,10 +13,10 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  firebase_core: ^2.24.2
-  firebase_auth: ^4.15.3
-  cloud_firestore: ^4.13.6
-  firebase_storage: ^11.5.6
+  firebase_core: ^3.8.1
+  firebase_auth: ^5.3.4
+  firebase_storage: ^12.3.7
+  cloud_firestore: ^5.6.0
   cached_network_image: ^3.3.0
   intl: ^0.19.0
   provider: ^6.1.1
@@ -21,11 +24,14 @@ dependencies:
   image_picker: ^1.0.5
   file_picker: ^6.1.1
   path_provider: ^2.1.1
+  path: ^1.8.3
+  permission_handler: ^11.1.0
   flutter_chat_bubble: ^2.0.2
   timeago: ^3.6.0
   flutter_linkify: ^6.0.0
   url_launcher: ^6.2.2
-  emoji_picker_flutter: ^1.6.3
+  emoji_picker_flutter: ^4.2.0
+  shimmer: ^3.0.0
 
 dev_dependencies:
   flutter_test:
@@ -41,3 +47,17 @@ flutter:
   assets:
     - assets/images/
     - assets/icons/
+
+topics:
+  - chat
+  - firebase
+  - messaging
+  - realtime
+  - group-chat
+
+platforms:
+  android:
+  ios:
+
+false_secrets:
+  - /example/**/google-services.json
