@@ -1,26 +1,14 @@
+import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:firebase_core/firebase_core.dart';
-import 'package:provider/provider.dart';
-import 'package:chatverse/chatverse.dart';
+import 'screens/login_screen.dart';
 import 'screens/chat_list_screen.dart';
+import 'package:firebase_auth/firebase_auth.dart' as auth show User;
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
-  try {
-    await Firebase.initializeApp();
-    runApp(const MyApp());
-  } catch (e) {
-    debugPrint('Failed to initialize Firebase: $e');
-    runApp(
-      MaterialApp(
-        home: Scaffold(
-          body: Center(
-            child: Text('Failed to initialize app: $e'),
-          ),
-        ),
-      ),
-    );
-  }
+  await Firebase.initializeApp();
+  runApp(const MyApp());
 }
 
 class MyApp extends StatelessWidget {
@@ -29,133 +17,39 @@ class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      title: 'ChatVerse Demo',
+      title: 'ChatVerse Example',
       theme: ThemeData(
-        colorScheme: ColorScheme.fromSeed(
-          seedColor: Colors.blue,
-          brightness: Brightness.light,
-        ),
+        primarySwatch: Colors.blue,
         useMaterial3: true,
-        appBarTheme: const AppBarTheme(
-          centerTitle: true,
-          elevation: 0,
-        ),
-        elevatedButtonTheme: ElevatedButtonThemeData(
-          style: ElevatedButton.styleFrom(
-            elevation: 0,
-            padding: const EdgeInsets.symmetric(
-              horizontal: 24,
-              vertical: 12,
-            ),
-            shape: RoundedRectangleBorder(
-              borderRadius: BorderRadius.circular(12),
-            ),
-          ),
-        ),
-        inputDecorationTheme: InputDecorationTheme(
-          border: OutlineInputBorder(
-            borderRadius: BorderRadius.circular(12),
-          ),
-          contentPadding: const EdgeInsets.symmetric(
-            horizontal: 16,
-            vertical: 16,
-          ),
-        ),
       ),
-      darkTheme: ThemeData(
-        colorScheme: ColorScheme.fromSeed(
-          seedColor: Colors.blue,
-          brightness: Brightness.dark,
-        ),
-        useMaterial3: true,
-        appBarTheme: const AppBarTheme(
-          centerTitle: true,
-          elevation: 0,
-        ),
-        elevatedButtonTheme: ElevatedButtonThemeData(
-          style: ElevatedButton.styleFrom(
-            elevation: 0,
-            padding: const EdgeInsets.symmetric(
-              horizontal: 24,
-              vertical: 12,
-            ),
-            shape: RoundedRectangleBorder(
-              borderRadius: BorderRadius.circular(12),
-            ),
-          ),
-        ),
-        inputDecorationTheme: InputDecorationTheme(
-          border: OutlineInputBorder(
-            borderRadius: BorderRadius.circular(12),
-          ),
-          contentPadding: const EdgeInsets.symmetric(
-            horizontal: 16,
-            vertical: 16,
-          ),
-        ),
-      ),
-      home: const ChatDemo(),
-    );
-  }
-}
+      home: StreamBuilder<auth.User?>(
+        stream: FirebaseAuth.instance.authStateChanges(),
+        builder: (context, snapshot) {
+          if (snapshot.connectionState == ConnectionState.waiting) {
+            return const Scaffold(
+              body: Center(
+                child: CircularProgressIndicator(),
+              ),
+            );
+          }
 
-class ChatDemo extends StatefulWidget {
-  const ChatDemo({Key? key}) : super(key: key);
+          if (snapshot.hasError) {
+            return Scaffold(
+              body: Center(
+                child: Text('Error: ${snapshot.error}'),
+              ),
+            );
+          }
 
-  @override
-  State<ChatDemo> createState() => _ChatDemoState();
-}
+          final user = snapshot.data;
+          print(user);
+          if (user != null) {
+            return ChatListScreen(userId: user.uid);
+          }
 
-class _ChatDemoState extends State<ChatDemo> {
-  late ChatController _chatController;
-  final String _userId = 'user1';
-  final Map<String, ChatUser> _users = {
-    'user1': ChatUser(
-      id: 'user1',
-      name: 'John Doe',
-      photoUrl: 'https://i.pravatar.cc/150?img=1',
-    ),
-    'user2': ChatUser(
-      id: 'user2',
-      name: 'Jane Smith',
-      photoUrl: 'https://i.pravatar.cc/150?img=2',
-    ),
-    'user3': ChatUser(
-      id: 'user3',
-      name: 'Mike Johnson',
-      photoUrl: 'https://i.pravatar.cc/150?img=3',
-    ),
-    'user4': ChatUser(
-      id: 'user4',
-      name: 'Sarah Wilson',
-      photoUrl: 'https://i.pravatar.cc/150?img=4',
-    ),
-  };
-
-  @override
-  void initState() {
-    super.initState();
-    _chatController = ChatController(
-      userId: _userId,
-      chatService: ChatService(userId: _userId),
-      authService: AuthService(),
-    );
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return ChangeNotifierProvider.value(
-      value: _chatController,
-      child: ChatListScreen(
-        users: _users,
-        currentUserId: _userId,
+          return const LoginScreen();
+        },
       ),
     );
-  }
-
-  @override
-  void dispose() {
-    _chatController.dispose();
-    super.dispose();
   }
 }
