@@ -1,18 +1,19 @@
 library chatverse;
 
-export 'src/models/message.dart';
-export 'src/models/chat_user.dart';
+export 'src/chat_controller.dart';
 export 'src/models/chat_room.dart';
-export 'src/widgets/chat_view.dart';
-export 'src/widgets/chat_input.dart';
-export 'src/widgets/chat_bubble.dart';
+export 'src/models/chat_user.dart';
+export 'src/models/message.dart';
 export 'src/services/chat_service.dart';
 export 'src/services/auth_service.dart';
+export 'src/widgets/chat_input.dart';
+export 'src/widgets/chat_list_view.dart';
+export 'src/widgets/profile_view.dart';
+export 'src/widgets/chat_view.dart';
+export 'src/widgets/new_chat_view.dart';
+export 'src/widgets/new_chat_fab.dart';
+export 'src/widgets/chat_room_tile.dart';
+export 'src/widgets/chat_image_uploader.dart';
+export 'src/widgets/user_details_widget.dart';
 export 'src/utils/chat_theme.dart';
-export 'src/chat_controller.dart';
 
-/// A Calculator.
-class Calculator {
-  /// Returns [value] plus 1.
-  int addOne(int value) => value + 1;
-}
