@@ -15,10 +15,14 @@ class Message {
   final String id;
   final String roomId;
   final String senderId;
+  final String senderName;
   final String content;
   final MessageType type;
   final DateTime createdAt;
   final String? replyTo;
+  final String? replyToMessageId;
+  final bool isRead;
+  final bool isDelivered;
   final Map<String, dynamic>? metadata;
   final Map<String, dynamic>? attachments;
 
@@ -26,10 +30,14 @@ class Message {
     required this.id,
     required this.roomId,
     required this.senderId,
+    required this.senderName,
     required this.content,
     required this.type,
     required this.createdAt,
     this.replyTo,
+    this.replyToMessageId,
+    this.isRead = false,
+    this.isDelivered = false,
     this.metadata,
     this.attachments,
   });
@@ -38,10 +46,14 @@ class Message {
     String? id,
     String? roomId,
     String? senderId,
+    String? senderName,
     String? content,
     MessageType? type,
     DateTime? createdAt,
     String? replyTo,
+    String? replyToMessageId,
+    bool? isRead,
+    bool? isDelivered,
     Map<String, dynamic>? metadata,
     Map<String, dynamic>? attachments,
   }) {
@@ -49,42 +61,66 @@ class Message {
       id: id ?? this.id,
       roomId: roomId ?? this.roomId,
       senderId: senderId ?? this.senderId,
+      senderName: senderName ?? this.senderName,
       content: content ?? this.content,
       type: type ?? this.type,
       createdAt: createdAt ?? this.createdAt,
       replyTo: replyTo ?? this.replyTo,
+      replyToMessageId: replyToMessageId ?? this.replyToMessageId,
+      isRead: isRead ?? this.isRead,
+      isDelivered: isDelivered ?? this.isDelivered,
       metadata: metadata ?? this.metadata,
       attachments: attachments ?? this.attachments,
     );
   }
 
-  Map<String, dynamic> toJson() {
+  Map<String, dynamic> toMap() {
     return {
       'id': id,
       'roomId': roomId,
       'senderId': senderId,
+      'senderName': senderName,
       'content': content,
       'type': type.toString().split('.').last,
-      'createdAt': createdAt.toIso8601String(),
+      'createdAt': Timestamp.fromDate(createdAt),
       if (replyTo != null) 'replyTo': replyTo,
+      if (replyToMessageId != null) 'replyToMessageId': replyToMessageId,
+      'isRead': isRead,
+      'isDelivered': isDelivered,
       if (metadata != null) 'metadata': metadata,
       if (attachments != null) 'attachments': attachments,
     };
   }
 
-  factory Message.fromJson(Map<String, dynamic> json) {
+  factory Message.fromMap(Map<String, dynamic> map) {
+    final createdAtValue = map['createdAt'];
+    final DateTime createdAt;
+    
+    if (createdAtValue is Timestamp) {
+      createdAt = createdAtValue.toDate();
+    } else if (createdAtValue is String) {
+      createdAt = DateTime.parse(createdAtValue);
+    } else {
+      createdAt = DateTime.now(); // Fallback
+    }
+
     return Message(
-      id: json['id'] as String,
-      roomId: json['roomId'] as String,
-      senderId: json['senderId'] as String,
-      content: json['content'] as String,
+      id: map['id']?.toString() ?? '',
+      roomId: map['roomId']?.toString() ?? '',
+      senderId: map['senderId']?.toString() ?? '',
+      senderName: map['senderName']?.toString() ?? 'Unknown User',
+      content: map['content']?.toString() ?? '',
       type: MessageType.values.firstWhere(
-        (e) => e.toString().split('.').last == json['type'],
+        (e) => e.toString().split('.').last == map['type']?.toString(),
+        orElse: () => MessageType.text,
       ),
-      createdAt: DateTime.parse(json['createdAt'] as String),
-      replyTo: json['replyTo'] as String?,
-      metadata: json['metadata'] as Map<String, dynamic>?,
-      attachments: json['attachments'] as Map<String, dynamic>?,
+      createdAt: createdAt,
+      replyTo: map['replyTo']?.toString(),
+      replyToMessageId: map['replyToMessageId']?.toString(),
+      isRead: map['isRead'] as bool? ?? false,
+      isDelivered: map['isDelivered'] as bool? ?? false,
+      metadata: map['metadata'] is Map ? Map<String, dynamic>.from(map['metadata']) : null,
+      attachments: map['attachments'] is Map ? Map<String, dynamic>.from(map['attachments']) : null,
     );
   }
 
@@ -95,10 +131,14 @@ class Message {
         other.id == id &&
         other.roomId == roomId &&
         other.senderId == senderId &&
+        other.senderName == senderName &&
         other.content == content &&
         other.type == type &&
         other.createdAt == createdAt &&
         other.replyTo == replyTo &&
+        other.replyToMessageId == replyToMessageId &&
+        other.isRead == isRead &&
+        other.isDelivered == isDelivered &&
         mapEquals(other.metadata, metadata) &&
         mapEquals(other.attachments, attachments);
   }
@@ -109,10 +149,14 @@ class Message {
       id,
       roomId,
       senderId,
+      senderName,
       content,
       type,
       createdAt,
       replyTo,
+      replyToMessageId,
+      isRead,
+      isDelivered,
       metadata,
       attachments,
     );
